Use master image as cache; simplify Makefile
Use the most recent image from the master branch as the build cache.
Pulling the current tag does not work for releases, because the tag
for a new release does not exist yet.

Pull the cache image in the build rule. This allows us to eliminate
the `pull` and `pullpush` rules, and to simplify the `build` rule and
the Travis configuration.