CI: Fix wrong cache image for release branches
The cache image uses the branch name as the version. Previously, we simply
hardcoded `master` as the branch name. This does not work anymore now that we
have release branches.

For tag builds, there is no general way to determine the branch from git or the
environment. However, we can determine the actual branch name from the git tag.
When we're on a release branch like `1.8`, TRAVIS_TAG will start with `v1.8.`.