CI: Do not upload cache images on release branches
For release branches, use the previous version on that branch as the cache
image. Like on `master`, we use the branch name for the cache image, relying on
the fact that images are tagged with an abbreviated version which matches the
release branch name, e.g. `1.8`.

Also, do not override VERSION unless we have to. We only need to override
VERSION when we are uploading a cache image on the master branch. Otherwise, we
can use the VERSION specified in the Makefile rather than deriving it from
TRAVIS_TAG.