Use the same cache image for PR and branch builds
- Pull branch image into Docker cache for PR builds

- Do not overwrite master cache image from PR builds

For PR builds, use the branch from which the PR originated, not the
target branch. This means that the same image is used for the Docker
build cache as for the corresponding branch build.

In Travis CI, the source branch is indicated by the
TRAVIS_PULL_REQUEST_BRANCH environment variable.